+++ PersistencePlugin/src/com/elmakers/mine/bukkit/plugins/persistence/PersistencePlugin.java
@@ -4,11 +4,11 @@ import java.io.File;
 import java.util.logging.Logger;
 
 import org.bukkit.Server;
-import org.bukkit.event.Event.Priority;
-import org.bukkit.event.Event.Type;
+import org.bukkit.command.Command;
+import org.bukkit.command.CommandSender;
+import org.bukkit.entity.Player;
 import org.bukkit.plugin.PluginDescriptionFile;
 import org.bukkit.plugin.PluginLoader;
-import org.bukkit.plugin.PluginManager;
 import org.bukkit.plugin.java.JavaPlugin;
 
 public class PersistencePlugin extends JavaPlugin
@@ -17,6 +17,18 @@ public class PersistencePlugin extends JavaPlugin
 	 * Public API
 	 */
 	
+	@Override
+	public boolean onCommand(CommandSender sender, Command cmd, String commandLabel, String[] args)
+	{
+		if (sender instanceof Player)
+		{
+			return handler.process((Player)sender, cmd, args);
+		}
+		
+		// TODO: Support server commands, etc.
+		return false;
+	}
+
 	public Persistence getPersistence()
 	{
 		persistence = Persistence.getInstance();
@@ -47,11 +59,8 @@ public class PersistencePlugin extends JavaPlugin
 
 	public void onEnable()
 	{
-		PluginManager pm = getServer().getPluginManager();
-		playerListener.initialize(this, persistence);
+		initialize();
 		
-        pm.registerEvent(Type.PLAYER_COMMAND, playerListener, Priority.Normal, this);
-        
 		PluginDescriptionFile pdfFile = this.getDescription();
         log.info(pdfFile.getName() + " version " + pdfFile.getVersion() + " is enabled");
 	}
@@ -72,11 +81,20 @@ public class PersistencePlugin extends JavaPlugin
 	}	
 	
 	/*
+	 * Internal functions
+	 */
+	
+	public void initialize()
+	{
+		handler.initialize(this, getPersistence());
+	}
+	
+	/*
 	 * Private data
 	 */
 	
 	private static PersistencePlugin pluginInstance = null;
-	private final PersistencePlayerListener playerListener = new PersistencePlayerListener();
+	private final PersistenceCommands handler = new PersistenceCommands();
 	private Persistence persistence = null;
 	private static final Logger log = Logger.getLogger("Minecraft");